requirements/base.txt
djangorestframework>=3.10.0
oauthlib==2.0.7
